Changed documentation to make it clearer, fixed typing, made a function in the pipeline that runs the pipeline. Changed and clean up of imports
